Added Authorization in server and improved error handling

+++ server/controllers/board.js
@@ -8,13 +8,17 @@ exports.createBoard = async (req, res) => {
     const { id } = req.params
     const project = await Project.findOne({ _id: id })
 
+    
     if(project) {
+      if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
       const { title } = req.body
       const board = new Board({
         title,
         projectId: id,
         tasks: []
       })
+
+      board.createdBy = req.user
 
       await board.save()
       project.boards.push(board)
@@ -23,9 +27,10 @@ exports.createBoard = async (req, res) => {
       return res.status(201).json(board)
     }
 
-    return res.sendStatus(404)
+    return res.status(404).json({ message: 'Project does not exist' })
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
 
@@ -37,15 +42,17 @@ exports.updateBoard = async (req, res) => {
     const board = await Board.findOne({ _id: boardId })
 
     if(board && (board.projectId === id)) {
+      if (req.user.id !== board.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
       board.title = title
       await board.save()
 
       return res.status(200).json(board)
     } else {
-      return res.sendStatus(404)
+      return res.status(404).json({ message: 'Board does not exist' })
     }
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
 
@@ -54,8 +61,11 @@ exports.deleteBoard = async (req, res) => {
 
   try {
     const project = await Project.findOne({ _id: id })
+    if (!project) return res.status(404).json({ message: 'Project does not exist' })
 
     if(project && project.boards.includes(mongoose.Types.ObjectId(boardId))) {
+      if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
+
       await Task.deleteMany({ projectId: id, boardId })
       await Board.findOneAndDelete({ _id: boardId })
       project.boards.pull(mongoose.Types.ObjectId(boardId))
@@ -63,10 +73,11 @@ exports.deleteBoard = async (req, res) => {
 
       return res.sendStatus(200)
     } else {
-      return res.sendStatus(404)
+      return res.status(404).json({ message: 'Board does not exist' })
     }
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
 
@@ -83,6 +94,8 @@ exports.moveTask = async (req, res) => {
     if(project 
       && moveFromBoard 
       && moveToBoard ) {
+        if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
+
         task.boardId = moveToBoardId
         moveFromBoard.tasks.pull(mongoose.Types.ObjectId(taskId))
         moveToBoard.tasks.push({
@@ -94,8 +107,11 @@ exports.moveTask = async (req, res) => {
         await moveFromBoard.save()
         await moveToBoard.save()
         res.sendStatus(200)
+      } else {
+        return res.status(404).json({ message: 'Resources does not exist' })
       }
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
