+++ server/controllers/project.js
@@ -4,7 +4,7 @@ const Task = require('../models/task')
 
 exports.getAllProjects = async (req, res) => {
   try {
-    const projects = await Project.find({})
+    const projects = await Project.find({ createdBy: req.user })
     return res.json({ projects })
   } catch (err) {
     console.log(err)
@@ -19,11 +19,15 @@ exports.getProject = async (req, res) => {
       populate: { path: 'tasks' }
     })
 
-    if (project) return res.json(project)
+    if (!project) return res.status(404).json({ message: 'Project does not exist' })
 
-    return res.sendStatus(404)
+    if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
+
+
+    return res.json(project)
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
 
@@ -31,6 +35,7 @@ exports.createProject = async (req, res) => {
   try {
     const { name } = req.body
     const project = new Project({ name, boards: [] })
+    project.createdBy = req.user
 
     await project.save()
 
@@ -44,13 +49,20 @@ exports.deleteProject = async (req, res) => {
   const { id } = req.params
 
   try {
+    const project = await Project.findOne({ _id: id })
+
+    if (!project) return res.status(404).json({ message: 'Project does not exist' })
+
+    if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
+
     await Task.deleteMany({ projectId: id })
     await Board.deleteMany({ projectId: id })
     await Project.findOneAndDelete({ _id: id })
-    
+
     return res.sendStatus(200)
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
 
@@ -60,15 +72,17 @@ exports.updateProject = async (req, res) => {
 
   try {
     const project = await Project.findOne({ _id: id })
-    if(project) {
-      project.name = name
-      await project.save()
 
-      return res.sendStatus(200)
-    } else {
-      return res.sendStatus(404)
-    }
+    if (!project) return res.status(404).json({ message: 'Project does not exist' })
+
+    if (req.user.id !== project.createdBy.toString()) return res.status(403).json({ message: 'Unauthorized to access this resource' })
+
+    project.name = name
+    await project.save()
+
+    return res.sendStatus(200)
   } catch (err) {
     console.log(err)
+    return res.status(500).json({ message: 'Some error has occurred please try again' })
   }
 }
